LayoutTests/inspector/editor/text-editor-type-over-decoration-expected.txt

This test checks that typing over the decoration does not collect decoration text as a part of damaged DOM.

Chunk model
Chunk [0] 0:10 (expanded)
Editor DOM
[0]       <div class="webkit-line-content">0</div>
[1]       <div class="webkit-line-content">1</div>
[2]       <div class="webkit-line-content">2</div>
[3]       <div class="webkit-line-content">3</div>
[4]       <div class="webkit-line-content">4</div>
[5:6]     <div class="webkit-line-content">5<div class="webkit-line-decorations"><span>Decoration content</span></div></div>
[6]       <div class="webkit-line-content">6</div>
[7]       <div class="webkit-line-content">7</div>
[8]       <div class="webkit-line-content">8</div>
[9]       <div class="webkit-line-content">9</div>
Editor DOM
[0]       <div class="webkit-line-content">0</div>
[1]       <div class="webkit-line-content">1</div>
[2]       <div class="webkit-line-content">2</div>
[3]       <div class="webkit-line-content">3</div>
[4]       <div class="webkit-line-content">4a5</div>
[5]       <div class="webkit-line-content">6</div>
[6]       <div class="webkit-line-content">7</div>
[7]       <div class="webkit-line-content">8</div>
[8]       <div class="webkit-line-content">9</div>
